Move RunSummaryFormatter into its own directory
There will be more run summary formatters coming.
